Renamed directory to kvstore
