+++ py-utils/setup.py
@@ -55,7 +55,7 @@ setup(name='cortx-py-utils',
                 'cortx.utils.ha.dm.repository',
                 'cortx.utils.ha',
                 'cortx.utils.validator',
-                'cortx.utils.store',
+                'cortx.utils.kvstore',
                 'cortx.utils.message_bus','cortx.utils.message_bus.tcp',
                 'cortx.utils.message_bus.tcp.kafka', 'cortx.utils.product_features',
                 'cortx.utils.security', 'cortx.utils.schema',
